refactor(hook): 쉼터 목록 로직을 useShelters 커스텀 훅으로 분리
